feat: resolve some AI agent frictions

+++ core/blong-browser/src/components/Card/Card.tsx
@@ -169,7 +169,7 @@ function resolveWidgetType(fieldSchema: IEnrichedFieldSchema, fieldName: string)
     if (fieldName.endsWith('Description')) return 'textArea';
     if (fieldSchema.type === 'integer') return 'integer';
     if (fieldSchema.type === 'number') return 'number';
-    if (fieldSchema.type === 'bigint') return 'integer';
+    if (fieldSchema.type === 'bigint') return 'bigint';
     if (fieldSchema.anyOf instanceof Array)
         return resolveWidgetType(
             fieldSchema.anyOf.find(schema => schema.type !== 'null') || fieldSchema.anyOf[0],
